+++ ludwig/explain/gbm.py
@@ -1,6 +1,6 @@
 from typing import List, Tuple
 
-from ludwig.explain.base import Explainer
+from ludwig.explain.explainer import Explainer
 from ludwig.explain.explanation import Explanation
 from ludwig.models.gbm import GBM
